helpers: file and rename functions. README.md added
